test(k-series): 退役腐烂的 test_kimi_live，K0 抽存为网关实网探针
0911 深度验证实证：test_kimi_live 的 K1–K6 依赖 AgentService.
send_turn/_worker_manager/_broker/_team_coordinator——该架构已被
Pi 单链取代（属性级 AttributeError），integration 标记使其休眠
腐烂不可见。逐一核实存活面后：
- 退役 K1–K6（对已删架构的断言无法'修复'——重写即重复既有
  真实模型门禁）。继任真实覆盖：agent 层 L08（真实 K3 诚实拒绝
  +fake REAL 零 action_txn）+ ADR-0007 授权链 + Gate 2 + A/B。
- K0 不依赖已删架构（纯网关层）——抽存为
  test_kimi_gateway_live.py，真实 K3 复验通过（11.9s：reachable/
  chat/strict tool/parallel calls/tool-result 回填）。
- AGENTD.md 测试入口与 W12 验收矩阵同步更新。

红→绿：红=send_turn AttributeError（腐烂实证）；绿=K0 真实 K3
通过 + 套件不再含引用已删符号的休眠腐烂。
